jupyter: factor out the kernel usage code into a separate hook (much better code modularity); unfortunately, I can't really test this, since the backend code is evidently mostly broken (see https://github.com/sagemathinc/cocalc/issues/5826)
